mierdacas
git-svn-id: http://pragma-studios.googlecode.com/svn/trunk@89 a041a8fe-e45c-11de-8d18-5b59c22968bc

+++ src/pragma/system/clock.cpp
@@ -7,6 +7,6 @@ namespace pragma
 {
 	TMillisecond GetTimeMilliseconds()
 	{
-		return (clock() * 1000 / CLOCKS_PER_SEC);
+		return ((float)clock() * 1000.f / CLOCKS_PER_SEC);
 	}
 }
